requirements.txt
genanki
langchain
langchain-community
langchain-experimental
langchain-openai
langchain-weaviate
langsmith
pydantic[dotenv]
pydantic-settings
pypdf
pytest
python-docx
python-dotenv
setuptools
srt
streamlit
tiktoken
tqdm
weaviate-client